+++ src/main/kotlin/GeneratorSettings.kt
@@ -1,17 +1,18 @@
 package ar.edu.itba.ss
 
+import java.math.BigDecimal
 import kotlin.random.Random
 
 data class GeneratorSettings(
     val random: Random,
     // Params
     val numberOfParticles: Int,
-    val radius: Double,
-    val mass: Double,
-    val initialVelocity: Double,
+    val radius: BigDecimal,
+    val mass: BigDecimal,
+    val initialVelocity: BigDecimal,
     val seed: Long,
     // System configuration
-    val obstacleRadius: Double,
-    val obstacleMass: Double?,
-    val containerRadius: Double
+    val obstacleRadius: BigDecimal,
+    val obstacleMass: BigDecimal?,
+    val containerRadius: BigDecimal
 )
