+++ src/main/kotlin/Settings.kt
@@ -1,12 +1,13 @@
 package ar.edu.itba.ss
 
 import java.io.File
+import java.math.BigDecimal
 
 data class Settings(
     val outputFile: File,
     val generatorSettings: GeneratorSettings,
     val particles: Map<Int, Particle>,
-    val finalTime: Double,
+    val finalTime: BigDecimal,
     val internalCollisions: Boolean,
     val eventDensity: Int?
 )
